refactor: 폴더 구조 변경. entities 에 있던것들 features 와 widget 으로 재구조
