+++ src/entities/comments/index.ts
@@ -1,0 +1,3 @@
+export * from "./api"
+export * from "./ui"
+export * from "./types"
