+++ src/entities/posts/api/postsApis.ts
@@ -1,27 +1,5 @@
 import { api } from "@/shared/api/api"
-
-// Types aligned with PostsManagerPage expectations
-export type Reaction = { likes: number; dislikes: number }
-
-export interface PostsResponse {
-  limit: number
-  posts: Post[]
-  total: number
-  skip: number
-}
-
-export type Post = {
-  id: number
-  body: string
-  title: string
-  userId: number
-  views: number
-  tags: string[]
-  reactions: Reaction
-}
-
-export type Tag = { url: string; slug: string; name: string }
-export type TagsResponse = Tag[]
+import { PostsResponse, TagsResponse, Post } from "../types"
 
 // API 함수
