+++ src/entities/users/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useUserActions } from "./useUserActions"
