+++ src/features/comments/index.ts
@@ -1,0 +1,2 @@
+export * from "./ui"
+export * from "./hooks"
