+++ src/features/comments/ui/index.ts
@@ -1,0 +1,3 @@
+export * from "./AddCommentDialog"
+export * from "./EditCommentDialog"
+export * from "./Comments"
