+++ src/features/posts/hooks/index.ts
@@ -1,1 +1,2 @@
 export { usePostActions } from "./usePostActions"
+export { useTagActions } from "./useTagActions"
