+++ src/features/posts/types/types.ts
@@ -1,0 +1,22 @@
+import { Post } from "@/entities/posts"
+import { User } from "@/entities/users"
+
+export interface CreatePostData {
+  title: string
+  body: string
+  userId: number
+}
+
+export interface PostWithAuthor extends Post {
+  author?: User
+}
+
+export interface PostsQueryResult {
+  posts: PostWithAuthor[]
+  total: number
+}
+
+export interface MutationContext {
+  previousPosts?: PostsQueryResult
+  tempId?: number
+}
