+++ src/widgets/index.ts
@@ -1,1 +1,2 @@
 export { Layout } from "./Layout"
+export * from "./posts"
