+++ src/widgets/posts/ui/PostsManagerComponent.tsx
@@ -1,15 +1,16 @@
 import { useState, useCallback } from "react"
 import { Plus } from "lucide-react"
 import { Button, Card, CardContent, CardHeader, CardTitle, useQueryParams } from "@/shared"
-import { PostsTable } from "../posts/ui/PostsTable"
-import { SearchFilterContainer } from "./SearchFilterContainer"
-import { Paginations } from "./Paginations"
-import { AddPostDialog, CreatePostData } from "./AddPostDialog"
-import { EditPostDialog, PostEditData } from "./EditPostDialog"
+import { PostsTable } from "@/features/posts/ui/PostsTable"
+import { SearchFilterContainer } from "../../../features/posts/ui/SearchFilterContainer"
+import { Paginations } from "../../../features/posts/ui/Paginations"
+import { AddPostDialog } from "../../../features/posts/ui/AddPostDialog"
+import { EditPostDialog, PostEditData } from "../../../features/posts/ui/EditPostDialog"
 import { PostDetailDialog } from "./PostDetailDialog"
-import { UserDialog } from "../users/ui/UserDialog"
-import { usePostActions, PostWithAuthor } from "./hooks/usePostActions"
-import { useUserActions } from "../users/hooks/useUserActions"
+import { UserDialog } from "@/entities/users/ui/UserDialog"
+import { usePostActions } from "@/features/posts/hooks/usePostActions"
+import { CreatePostData, PostWithAuthor } from "@/features/posts/types"
+import { useUserActions } from "@/entities/users/hooks/useUserActions"
 
 export const PostsManagerComponent = () => {
   const { skip, limit, searchQuery, selectedTag, sortBy, order } = useQueryParams()
